Fix apk files not removed from /data/app in some cases

+++ misc/module_template/AppUpdate/customize.sh
@@ -17,10 +17,12 @@ appendToScript() {
 
 removeDataApp() {
 	result=$(pm path "$1")
-	path=$(dirname $(dirname ${result:8}))
-	if [ ${path:0:10} = "/data/app/" ]; then
-		ui_print "- Update app: $1, remove on next boot: $path"
-		appendToScript "rm -rf $path"
+	if [ -n "$result" ]; then
+		path=$(dirname $(dirname ${result:8}))
+		if [ ${path:0:10} = "/data/app/" ]; then
+			ui_print "- Update app: $1, remove on next boot: $path"
+			appendToScript "rm -rf $path"
+		fi
 	fi
 }
